test: type on-book prices as u128 to drop the `as u128` casts
Following up on the u128-PR review: the test price locals are now `u128`, so
`price * PRICE_SCALE` no longer needs a cast. To keep the expected-value
arithmetic clean (`price * quantity`, `balance(price * lot, …)`), `Quantity`
gains a `From<u128>` impl and the `balance` test helper takes `impl
Into<Quantity>`. Because two `From<int>` impls make bare integer literals
ambiguous, literal arguments feeding `Into<Quantity>` now carry explicit `u64`
suffixes.

DEFI-2848

+++ canister/src/dashboard/tests.rs
@@ -213,13 +213,13 @@ fn place(
     state: &mut State<VectorMemory, VectorMemory>,
     user: Principal,
     side: Side,
-    price: u64,
+    price: u128,
     quantity: u64,
 ) -> OrderId {
     let pair: TradingPair = icp_ckbtc_trading_pair();
     let pending = PendingOrder {
         side,
-        price: Price::new(price as u128 * PRICE_SCALE),
+        price: Price::new(price * PRICE_SCALE),
         quantity: Quantity::from(quantity),
     };
     let (token, required) = match pending.side {
